pdf2text2chunks/requirements.txt
# PDF to RAG Converter Requirements

# Core dependencies
PyPDF2>=3.0.0
openai>=1.0.0
spacy>=3.4.0
tqdm>=4.64.0
pathlib2>=2.3.7; python_version < '3.4'
rank_bm25>=0.2.2  # Added for BM25

# Azure services (optional)
azure-ai-translation-text>=1.0.0
azure-storage-blob>=12.0.0
azure-core>=1.26.0

# Google Cloud services (optional)
google-cloud-vision>=3.0.0
google-cloud-translate>=3.8.0
google-cloud-storage>=2.7.0

# OCR and image processing (optional)
PyMuPDF>=1.20.0  # fitz
opencv-python>=4.6.0
Pillow>=9.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
easyocr>=1.6.0

# Text processing
arabic-reshaper>=2.1.3
python-bidi>=0.4.2

# System utilities
psutil>=5.9.0
python-dotenv>=0.19.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950